app/requirements.txt
Flask
gunicorn
Pytest
psycopg2-binary